+++ CHANGELOG.md
@@ -8,6 +8,7 @@
   * `send` has been renamed to `push` to unify client and server messaging commands
 
 * Backwards incompatible changes
+  * [Generator] `mix phoenix.gen.resource` renamed to `mix phoenix.gen.html`
   * [Channel] `reply` has been renamed to `push` to better signify we are only push a message down the socket, not replying to a specific request
   * [HTML] `use Phoenix.HTML` no longer imports controller functions. You must add `import Phoenix.Controller, only: [get_flash: 2]` manually to your views or your `web.ex`
   * [Endpoint] Code reloader must now be configured in your endpoint instead of Phoenix. Therefore, upgrade your `config/dev.exs` replacing